Added file alredy exists check when pasting

+++ actiones.py
@@ -13,9 +13,13 @@ class actiones():
 
     def paste(self,window_command,path):
         if self.clipbord:
-            shutil.copy2(self.clipbord,path)
-            window_command.addstr(0,0,'Pasted',curses.A_REVERSE)
-            window_command.refresh()
+            if not os.path.exists(os.path.join(path,os.path.basename(self.clipbord))):
+                shutil.copy2(self.clipbord,path)
+                window_command.addstr(0,0,'Pasted',curses.A_REVERSE)
+            else:
+                window_command.addstr(0,0,'Item already exists',curses.A_REVERSE)
+        window_command.addstr(0,0,'No path in clipbord',curses.A_REVERSE)
+        window_command.refresh()
 
     def delete(self,window_command,path,stdscr):
         if path:
